+++ litecoin_scrypt/scryptmodule.c
@@ -1,16 +1,22 @@
 #include <Python.h>
 #include "scrypt.h"
 
-/* Python 2.x version */
+/* Python 2.x / PyPy binding */
 static PyObject *scrypt_getpowhash(PyObject *self, PyObject *args)
 {
     char *output;
     PyObject *value;
     PyStringObject *input;
+    (void)self;  /* required by Python C API convention */
+
     if (!PyArg_ParseTuple(args, "S", &input))
         return NULL;
     Py_INCREF(input);
     output = PyMem_Malloc(32);
+    if (output == NULL) {
+        Py_DECREF(input);
+        return PyErr_NoMemory();
+    }
 
     scrypt_1024_1_1_256((char *)PyString_AsString((PyObject*) input), output);
     Py_DECREF(input);
@@ -24,6 +30,7 @@ static PyMethodDef ScryptMethods[] = {
     { NULL, NULL, 0, NULL }
 };
 
+PyMODINIT_FUNC initltc_scrypt(void);
 PyMODINIT_FUNC initltc_scrypt(void) {
     (void) Py_InitModule("ltc_scrypt", ScryptMethods);
 }